Move utils to lib dir, remove goatcounter
