--- hooks/useQueryParam.tsx
@@ -1,18 +1,0 @@
-import { GetServerSidePropsContext } from "next";
-import { NextRouter } from "next/router";
-
-export function useQueryParam(
-  contextOrRouter: GetServerSidePropsContext | NextRouter,
-  name: string,
-  fallback: any
-) {
-  const { query } = contextOrRouter;
-  if (!query[name]) {
-    return fallback;
-  }
-  const param = query[name];
-  if (param instanceof Array) {
-    return param[0];
-  }
-  return param;
-}
