+++ pages/legal.tsx
@@ -4,7 +4,7 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import styled from "styled-components";
 import { PaddedSection } from "../components/PaddedSection";
 import SEO from "../components/SEO";
-import { generateRSSFeeds } from "../utils/generateRSSFeeds";
+import { generateRSSFeeds } from "../lib/generateRSSFeeds";
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => {
   // This is an ugly way to generate the RSS feeds during static generation
